feat: update chatbot examples based on review feedback
